bitcoin-model/requirements.txt
uvicorn
fastapi
tweepy==4.10.1
transformers
python-dotenv
scipy
pyyaml
pandas
torch
numpy
tritonclient[http]
sse_starlette
jinja2